refactor(scales): move the MiBeacon frame helpers into a shared module
Lift decryptMiBeaconV5, macFrameOrderFromFrame, normUuid and the FE95
constants out of xiaomi-s800.ts into scales/mibeacon.ts, unchanged, so a
second MiBeacon scale can reuse them. The S800 module re-exports the moved
functions, so its public surface and existing tests are untouched.
